requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
attrs==23.2.0
black==24.3.0
certifi==2024.2.2
click==8.1.7
flake8==7.0.0
frozenlist==1.4.1
greenlet==3.0.3
idna==3.6
isort==5.13.2
magic-filter==1.0.12
mccabe==0.7.0
multidict==6.0.5
mypy==1.9.0
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
pycodestyle==2.11.1
pydantic==2.5.3
pydantic_core==2.14.6
pyflakes==3.2.0
python-dotenv==1.0.1
PyYAML==6.0.1
SQLAlchemy==2.0.28
types-PyYAML==6.0.12.20240311
typing_extensions==4.10.0
yarl==1.9.4